+++ apps/angular-i18next-demo/src/app/features/rich_form_feature/rich-form.model.ts
@@ -1,0 +1,11 @@
+export class Contact {
+    lastName: string | undefined;
+    firstName: string | undefined;
+    middleName: string | undefined;
+}
+
+export class RichFormModel {
+    email: string | undefined;
+    count = 0;
+    technicalContact: Contact = new Contact();
+}