+++ libs/angular-i18next/src/lib/I18NextEvents.ts
@@ -5,10 +5,11 @@ import {
   MissingKeyEvent,
   ResourceEvent,
 } from './ITranslationEvents';
+import * as i18n from 'i18next';
 
 export class I18NextEvents implements ITranslationEvents {
-  initialized = new BehaviorSubject(false);
-  loaded = new BehaviorSubject({});
+  initialized = new BehaviorSubject<i18n.InitOptions | undefined>(undefined);
+  loaded = new BehaviorSubject(false);
   failedLoading = new Subject();
   missingKey = new Subject<MissingKeyEvent>();
   added = new Subject<ResourceEvent>();