+++ libs/angular-i18next/src/lib/ITranslationEvents.ts
@@ -1,11 +1,12 @@
 import { BehaviorSubject, Subject } from 'rxjs';
+import * as i18n from 'i18next';
 
 export type ResourceEvent = { lng: any; ns: any };
 export type MissingKeyEvent = { lngs: any; namespace: any; key: any; res: any };
 
 export interface ITranslationEvents {
-  initialized: BehaviorSubject<any>;
-  loaded: BehaviorSubject<{ [language: string]: readonly string[] }>;
+  initialized: BehaviorSubject<i18n.InitOptions | undefined>;
+  loaded: BehaviorSubject<boolean>;
   failedLoading: Subject<any>;
   missingKey: Subject<MissingKeyEvent>;
   added: Subject<ResourceEvent>;