+++ libs/angular-i18next/src/lib/models.ts
@@ -1,8 +1,10 @@
-import type { TOptions } from 'i18next';
+import * as i18n from 'i18next';
 
-export type FormatPipeOptions = { format?: string; lng?: string };
+export type FormatPipeOptions = { format?: string; lng?: string; case?: string; [key: string]: any };
 export type PrependPipeOptions = {
   prependScope?: boolean;
   prependNamespace?: boolean;
 };
-export type PipeOptions = TOptions & FormatPipeOptions & PrependPipeOptions;
+export type PipeOptions = i18n.TOptionsBase & i18n.StringMap & { defaultValue?: string; } &
+  FormatPipeOptions &
+  PrependPipeOptions;