+++ libs/angular-i18next/src/tests/pipes/I18NextEagerPipe.spec.ts
@@ -1,28 +1,37 @@
-import { ChangeDetectorRef } from '@angular/core';
-import { ITranslationEvents, PipeOptions } from '../../lib/index';
-import { I18NextEagerPipe } from '../../lib/I18NextEagerPipe';
-import { BehaviorSubject } from 'rxjs';
-import { MockI18NextService } from '../mocks/MockTranslationService';
+import { ApplicationInitStatus, ChangeDetectorRef } from '@angular/core';
+import { TestBed } from '@angular/core/testing';
+import { I18NextEagerPipe, I18NextModule, I18NEXT_NAMESPACE, I18NEXT_SCOPE, I18NEXT_SERVICE, ITranslationService, PipeOptions } from '../../lib';
+import { I18N_PROVIDERS } from '../setup';
+
 
 describe('I18NextEagerPipe', () => {
   let pipe: I18NextEagerPipe,
-    languageChangedSubject: BehaviorSubject<string>,
-    changeDetector: ChangeDetectorRef,
-    service: MockI18NextService;
+    service: ITranslationService,
+    markForCheckSpy: jest.SpyInstance<any, unknown[]>;
 
-  beforeEach(() => {
-    service = new MockI18NextService();
-    service.language = 'en';
-    languageChangedSubject = new BehaviorSubject<string>('en');
+  beforeEach(async () => {
+      TestBed.configureTestingModule({
+        imports: [I18NextModule.forRoot()],
+        providers: [...I18N_PROVIDERS,
+          { provide: ChangeDetectorRef, useValue: { detectChanges: () => { }, markForCheck: () => { } } }, {
+            provide: I18NEXT_SCOPE,
+            useValue: 'scope'
+          },
+          {
+            provide: I18NEXT_NAMESPACE,
+            useValue: 'ns'
+          }],
+      });
 
-    const scope: string = 'scope';
-    const ns: string = 'ns';
-    service.events = {
-      languageChanged: languageChangedSubject,
-    } as ITranslationEvents;
+      // until https://github.com/angular/angular/issues/24218 is fixed
+      await TestBed.inject(ApplicationInitStatus).donePromise;
 
-    // changeDetector = jasmine.createSpyObj<ChangeDetectorRef>(['markForCheck']);
-    //pipe = new I18NextEagerPipe(service, ns, scope, changeDetector);
+      service = TestBed.inject(I18NEXT_SERVICE);
+      pipe = TestBed.inject(I18NextEagerPipe);
+      const changeDetector = TestBed.inject(ChangeDetectorRef);
+
+      // So, I am spying directly on the prototype.
+      markForCheckSpy = jest.spyOn(changeDetector, 'markForCheck');
   });
 
   it('should create the pipe', () => {
@@ -34,7 +43,7 @@ describe('I18NextEagerPipe', () => {
     let myOptions: PipeOptions;
 
     beforeEach(() => {
-      myOptions = { myValue: 'value1' };
+      myOptions = { defaultValue: 'value1' };
 
       result = pipe.transform('myKey', myOptions);
     });
@@ -42,7 +51,7 @@ describe('I18NextEagerPipe', () => {
     it('should call the translate service', () => {
       expect(service.t).toHaveBeenCalledWith(
         ['ns:scope.myKey', 'scope.myKey', 'ns:myKey', 'myKey'],
-        { myValue: 'value1' }
+        myOptions
       );
     });
 
@@ -52,24 +61,24 @@ describe('I18NextEagerPipe', () => {
 
     describe('when the language changes', () => {
       beforeEach(() => {
-        languageChangedSubject.next('es');
+        service.events.languageChanged.next('es');
         service.language = 'es';
       });
 
       it('should mark for check so it triggers the pipe transform', () => {
-        expect(changeDetector.markForCheck).toHaveBeenCalled();
+        expect(markForCheckSpy).toHaveBeenCalled();
       });
 
       describe('when the pipe gets triggered by change detection', () => {
         beforeEach(() => {
-          service.t.calls.reset();
+          jest.clearAllMocks();
           result = pipe.transform('myKey', myOptions);
         });
 
         it('should call the translate service', () => {
           expect(service.t).toHaveBeenCalledWith(
             ['ns:scope.myKey', 'scope.myKey', 'ns:myKey', 'myKey'],
-            { myValue: 'value1' }
+            { defaultValue: 'value1' }
           );
         });
 
@@ -81,7 +90,7 @@ describe('I18NextEagerPipe', () => {
 
     describe('when called with same key and options', () => {
       beforeEach(() => {
-        service.t.calls.reset();
+        jest.clearAllMocks();
         result = pipe.transform('myKey', myOptions);
       });
 
@@ -96,14 +105,14 @@ describe('I18NextEagerPipe', () => {
 
     describe('when called with same key but different options', () => {
       beforeEach(() => {
-        service.t.calls.reset();
-        result = pipe.transform('myKey', { myValue: 'value2' });
+        jest.clearAllMocks();
+        result = pipe.transform('myKey', { defaultValue: 'value2' });
       });
 
       it('should call the translate service', () => {
         expect(service.t).toHaveBeenCalledWith(
           ['ns:scope.myKey', 'scope.myKey', 'ns:myKey', 'myKey'],
-          { myValue: 'value2' }
+          { defaultValue: 'value2' }
         );
       });
 
@@ -114,14 +123,14 @@ describe('I18NextEagerPipe', () => {
 
     describe('when called with different key but same options', () => {
       beforeEach(() => {
-        service.t.calls.reset();
+        jest.clearAllMocks();
         result = pipe.transform('myKey2', myOptions);
       });
 
       it('should call the translate service', () => {
         expect(service.t).toHaveBeenCalledWith(
           ['ns:scope.myKey2', 'scope.myKey2', 'ns:myKey2', 'myKey2'],
-          { myValue: 'value1' }
+          { defaultValue: 'value1' }
         );
       });
 
@@ -151,7 +160,7 @@ describe('I18NextEagerPipe', () => {
 
     describe('when called with same key', () => {
       beforeEach(() => {
-        service.t.calls.reset();
+        jest.clearAllMocks();
         result = pipe.transform('myKey');
       });
 
@@ -166,7 +175,7 @@ describe('I18NextEagerPipe', () => {
 
     describe('when called with different key', () => {
       beforeEach(() => {
-        service.t.calls.reset();
+        jest.clearAllMocks();
         result = pipe.transform('myKey2');
       });
 
@@ -183,3 +192,4 @@ describe('I18NextEagerPipe', () => {
     });
   });
 });
+